requirements.txt
# Core PPT generation
python-pptx>=0.6.21

# LLM & Agent Framework
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# Hugging Face
transformers>=4.40.0
huggingface-hub>=0.22.0

# Image Generation (Stable Diffusion)
diffusers>=0.27.0
accelerate>=0.29.0

# PyTorch (CPU version - use cu118 index for GPU)
torch>=2.1.0
torchvision>=0.16.0

# HTTP & Environment
requests>=2.31.0
python-dotenv>=1.0.0

# Image handling
Pillow>=10.3.0

# CLI
rich>=13.7.0
click>=8.1.7

# Logging
colorlog>=6.8.2

# Async support
aiohttp>=3.9.0
aiofiles>=23.2.1